requirements.txt
imapclient
tqdm
python-dotenv
